# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "flock",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tools/": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Create Tool",
        "description": "Create new tool.",
        "operationId": "create_tool",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolBase"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Read Tools",
        "operationId": "read_tools",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{provider_id}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Read Tool",
        "operationId": "read_tool",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/{tool_id}": {
      "put": {
        "tags": [
          "tools"
        ],
        "summary": "Update Tool",
        "description": "Update a tool.",
        "operationId": "update_tool",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tools"
        ],
        "summary": "Delete Tool",
        "description": "Delete a tool.",
        "operationId": "delete_tool",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/toolprovider/": {
      "get": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Read Provider List With Tools",
        "operationId": "read_provider_list_with_tools",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersListWithToolsOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolProvider"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/{tool_provider_id}": {
      "get": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Read Provider",
        "description": "Get provider by ID.",
        "operationId": "read_provider",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Update Provider",
        "description": "Update a provider.",
        "operationId": "update_provider",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Delete Provider",
        "description": "删除工具提供者",
        "operationId": "delete_provider",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/withtools/{tool_provider_id}": {
      "get": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Read Provider With Tools",
        "operationId": "read_provider_with_tools",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolProviderWithToolsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/{tool_provider_id}/authenticate": {
      "post": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Authenticate Provider",
        "description": "对工具提供商进行鉴权或刷新工具列表",
        "operationId": "authenticate_provider",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/mcp": {
      "post": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Create Mcp Provider",
        "description": "创建 MCP 工具提供者并同步工具",
        "operationId": "create_mcp_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/mcp/{tool_provider_id}": {
      "put": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Update Mcp Provider",
        "description": "更新 MCP 工具提供者并同步工具",
        "operationId": "update_mcp_provider",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/mcp/test-connection": {
      "post": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Test Mcp Connection",
        "description": "测试 MCP 连接",
        "operationId": "test_mcp_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPConnectionTest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Toolprovider-Test Mcp Connection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolprovider/mcp/{tool_provider_id}/sync": {
      "post": {
        "tags": [
          "toolprovider"
        ],
        "summary": "Sync Mcp Tools",
        "description": "同步 MCP 工具",
        "operationId": "sync_mcp_tools",
        "parameters": [
          {
            "name": "tool_provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toolprovider-Sync Mcp Tools"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolCreate": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,64}$",
            "title": "Name",
            "unique": true
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "managed": {
            "type": "boolean",
            "title": "Managed",
            "default": false
          },
          "tool_definition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tool Definition"
          },
          "input_parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Parameters"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "is_online": {
            "type": "boolean",
            "title": "Is Online",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "tool_definition",
          "provider_id"
        ],
        "title": "ToolCreate"
      },
...
      "ToolsOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ToolOut"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ToolsOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}